+++ controlnet_ext/__init__.py
@@ -13,8 +13,12 @@ except ImportError:
         get_cn_models,
     )
 
+from .restore import CNHijackRestore, cn_allow_script_control
+
 __all__ = [
     "ControlNetExt",
+    "CNHijackRestore",
+    "cn_allow_script_control",
     "controlnet_exists",
     "controlnet_type",
     "get_cn_models",